Store vertex groups in a more GPU friendly way
Store an equal number of weights per vertex. The number
of weights stored per vertex will always accommodate
the vertex with the largest number of weights. For each
vertex weights are stored in decreasing making it easy
for a renderer to pick only the first N and get fair
quality.